+++ projects/slormancer-api/src/constants/content/data/data-mechanic.ts
@@ -82,7 +82,7 @@ export const DATA_MECHANIC: { [key: string]:  DataMechanic} = {
     [MechanicType.Blorm]: {
         values: [
             effectValueConstant(10, false, 'garbage_stat', EffectValueValueType.Stat),
-            effectValueSynergy(100, 0, EffectValueUpgradeType.None, false, 'basic_damage', 'basic_damage', EffectValueValueType.Stat, undefined, 3),
+            effectValueSynergy(100, 0, EffectValueUpgradeType.None, false, 'physical_damage', 'physical_damage', EffectValueValueType.Stat, undefined, 3),
             effectValueConstant(100, false, 'garbage_stat', EffectValueValueType.Stat),
         ],
         genres: [SkillGenre.Minion],