+++ CHANGELOG.md
@@ -55,6 +55,7 @@ Initial release
 - Reaper primordial data are no longer lost when parsed from a save file
 - Attributes now allow invalid values on save parsing to avoid issues with corrupted saves
 - Ancestral legacy nodes should no longer allow impossible configurations
+- Blorm damages is now correctly based on skill damage
 
 ## TODO
 - mise à jour packages